Adds integration test for bg task
